# mono/server-go/api/openapi.yaml
openapi: "3.0.3"
info:
  title: RTVortex API
  description: |
    RTVortex is an AI-powered code review engine. This API provides endpoints for
    managing repositories, triggering reviews, handling OAuth authentication,
    configuring LLM providers, and receiving webhook events from VCS platforms.
  version: "1.0.0"
  contact:
    name: AuralithAI
    url: https://github.com/AuralithAI/rtvortex
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.rtvortex.example.com
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from the OAuth flow.

  # ── Reusable schemas ─────────────────────────────────────────────────────
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Not Found"
        message:
          type: string
          example: "resource not found"
        status:
          type: integer
          example: 404
      required: [error, message, status]
...
    CompletionRequest:
      type: object
      properties:
        model:
          type: string
          description: Model identifier (provider-specific). Uses provider default if omitted.
          example: "gpt-4o"
        messages:
          type: array
          items:
            type: object
            properties:
              role:
                type: string
                enum: [system, user, assistant]
              content:
                type: string
            required: [role, content]
        max_tokens:
          type: integer
          example: 2048
        temperature:
          type: number
          format: double
          example: 0.7
        top_p:
          type: number
          format: double
          example: 1.0
        stop:
          type: array
          items:
            type: string
      required: [messages]
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: integer
            description: Unix timestamp when the window resets
        Retry-After:
          schema:
            type: integer
            description: Seconds until the client should retry
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              retry_after_seconds:
                type: integer

# ── Paths ────────────────────────────────────────────────────────────────────
paths:
...
  /api/v1/llm/stream:
    post:
      tags: [LLM]
      summary: Stream an LLM completion (SSE)
      description: |
        Sends a completion request and returns a Server-Sent Events stream.
        Each event is a JSON `StreamChunk` containing incremental content.
        The stream ends with a `data: [DONE]` sentinel.

        Supports OpenAI, Anthropic, and Ollama providers with native streaming.
        Non-streaming providers fall back to a single-chunk response.
      operationId: streamLLMCompletion
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompletionRequest"
      responses:
        "200":
          description: SSE stream of completion chunks
          content:
            text/event-stream:
              schema:
                type: string
                description: |
                  Newline-delimited SSE events:
                  `data: {"content":"Hello","done":false}`
                  `data: {"content":"","done":true,"finish_reason":"stop","usage":{...}}`
                  `data: [DONE]`
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "502":
          description: LLM provider error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── Admin ────────────────────────────────────────────────────────────────
...